refactor: empty이미지를 띄워주는 box 컴포넌트로 분리
