+++ frontend/src/components/EmptyBox/index.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Wrapper, Logo } from './style';
+import emptyImg from '../../../public/assets/empty.svg';
+
+const EmptyBox = () => {
+    return (
+        <Wrapper>
+            <Logo aria-label="empty" src={emptyImg} />
+        </Wrapper>
+    );
+};
+
+export default EmptyBox;
