+++ frontend/src/components/EmptyBox/style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+    width: 70%;
+    height: 100%;
+    margin: 20px auto;
+`;
+
+export const Logo = styled.img`
+    width: 100%;
+    height: 100%;
+`;
