+++ frontend/src/pages/Statistics/style.js
@@ -40,14 +40,3 @@ export const DonutBox = styled.div`
         border-radius: none;
     }
 `;
-
-export const EmptyBox = styled.div`
-    width: 70%;
-    height: 100%;
-    margin: 20px auto;
-`;
-
-export const Logo = styled.img`
-    width: 100%;
-    height: 100%;
-`;
